Added InitialRefactor migration, fixed Movie DTOs and added nullable props to entities

+++ DTOs/MovieDTO.cs
@@ -10,7 +10,7 @@ namespace dadachMovie.DTOs
         public string ReleaseDatePersian { get; set; }
         public float ImdbRate { get; set; }
         public int ImdbRatesCount { get; set; }
-        public float MetacriticRate { get; set; }
+        public int MetacriticRate { get; set; }
         public double AverageUserRate { get; set; }
         public int TotalUserRatesCount { get; set; }
         public int Lenght { get; set; }
